+++ lib/src/gmt_web.dart
@@ -3,20 +3,33 @@ import 'dart:async';
 import 'package:http/http.dart' as http;
 
 class GMTIpl {
-  static Future<String?> now() async {
+  static Future<String?> now({
+    List<String> urls = const [
+      'https://www.example.com',
+      'https://www.google.com',
+    ],
+    Duration? timeoutOfEach,
+  }) async {
     http.Response? response;
+
     try {
-      response = await http.get(Uri.base);
+      if (timeoutOfEach != null) {
+        response = await http.get(Uri.base).timeout(timeoutOfEach);
+      } else {
+        response = await http.get(Uri.base);
+      }
 
       if (response.statusCode != 200) {
+        response = null;
         throw Exception('GMT get statusCode != 200');
       }
+
+      print('[GMT] get from current URL successfully');
     } catch (e) {
-      print(
-          'GMT get from current URL error: $e\nGMT can\'t get DateTime from internet!!!');
+      print('[GMT] error: get from current URL error: $e');
     }
     String? dateTime = response?.headers['date'];
-    if (response != null && response.statusCode == 200 && dateTime != null) {
+    if (dateTime != null && response!.statusCode == 200) {
       return dateTime;
     } else {
       return null;