+++ test/gmt_test.dart
@@ -6,9 +6,12 @@ import 'package:test/test.dart';
 void main() {
   test('test GMT.now', () async {
     Timer.periodic(const Duration(seconds: 1), (timer) async {
-      print(await GMT.now());
+      print(await GMT.now(urls: [
+        'https://www.example.com',
+        'https://www.google.com',
+      ], timeoutOfEach: Duration(seconds: 1)));
     });
 
-    await Future.delayed(const Duration(seconds: 5));
+    await Future.delayed(const Duration(seconds: 10));
   });
 }